+++ completion/__init__.py
@@ -3,4 +3,4 @@ Completion App
 """
 
 
-__version__ = '4.4.1'
+__version__ = '4.5.0'